test(llm-shared): split real-sdk signature tests out of test_anthropic_provider.py (#15016)
test_anthropic_provider.py grew to 638 lines adding the #15016 signature-
binding tests, over the file-size guard's 600-line MAX_LINES with no
grandfather entry (and none may be added -- the ratchet only shrinks).
Moved TestRouteSamplingKwargs, TestThinkingBudgetToEffort and
TestKwargsMatchRealSdkSignature into a new sibling module,
test_anthropic_sampling_kwargs.py, grouped by the concern they actually
test (the SDK 1.x sampling-kwargs boundary) rather than by file-size
avoidance alone. Both files are now well under the limit.
